// budgetFlow-ui/src/app/services/models/expense-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoryRequest } from '../models/category-request';
export interface ExpenseRequest {
  amount?: number;
  categoryRequest?: CategoryRequest;
  expenseDate?: string;
  id?: number;
  name?: string;
  note?: string;
}